Create Virgo Restaurant's Info
This project showcases Virgo's menu items in many different perspectives, the first perspective is detailed information about the menu. However, once you get to the second part of the table, it will show which menu item is higher than $10. In this case, there are two items which fulfill the requirement of the specification; the chef platter and the grand slam. As for the last table, it consists of all the prices in order from least to greatest.